orgit-log-open: Call magit-log-arguments from repo
Calling orgit-log-open from an Org buffer that is not in a Git
repository signals a magit-outside-git-repo error that is triggered by
the call to magit-log-arguments.  Holding the Magit version constant
at the current master (0f1234e7), this failure bisects to
a2d3937 (Adapt to changes in Magit, 2019-04-17).

Reposition the default-directory binding so that magit-log-arguments
is called within the linked repo.

Note that orgit-rev-open calls magit-diff-arguments, so it _could_ hit
the same failure, but that call already happens with default-directory
correctly bound.

Fixes #34.