eval-after-load: replace string feature w/ symbol
TL;DR Given (eval-after-load FILE FORM), eval-after-load is less
efficient if FILE is a string than when it's a symbol.

When FILE is a string, eval-after-load creates a regexp entry in
after-load-alist that Emacs uses to scan the path of *every* file it
loads, and continues to do so after FILE is loaded.

When FILE is a symbol, you avoid this polling behavior. Instead, there
is no such aggressive lookup behavior. FORM isn't evaluated until
a (provide 'FILE) is evaluated in the target package.