Avoid loading Magit whenever Org is loaded
Magit was loaded because Orgit autoloaded calls to the autoloaded
`orgit-link-set-parameters'.  These calls are only evaluated after
Org is loaded but then it has to load Orgit to make that function
available and doing that loads Magit, which is slow.

Fix that by using `org-link-set-parameters' directly instead of our
wrapper `orgit-link-set-parameters'.  The latter was only necessary
to support Org versions before v9.0, but since we now depend on at
least Org v9.4 it no longer serves a purpose.